refactor: move regression line code into its own module
Extract the regression fitting and plotting helpers (add_regression_line
and its private helpers, constants, and positivity requirements) out of
graph_utils.py into a dedicated regression.py module, and relocate the
corresponding tests to test_regression.py.
